+++ lib/ar_captcha.dart
@@ -1,4 +1,5 @@
 library;
 
+export 'res/enums/data_size.dart';
 export 'res/enums/captcha_type.dart';
 export 'controller/ar_captcha_controller.dart';