+++ lib/res/enums/data_size.dart
@@ -1,0 +1,8 @@
+/// Check the checkBox automatic before captcha showed
+enum DataSize {
+  /// Show ar captcha checkbox section
+  normal,
+
+  /// Not show ar captcha checkbox section
+  invisible,
+}